pyproject.toml
[project]
name = "poisontr33s-dynamic-mcp"
version = "0.1.0"
description = "Dynamic MCP server orchestration with RAG/CAG capabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "PyYAML>=6.0.0",
    "GitPython>=3.1.0",
    "mcp>=1.0.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "C4", "PT", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"